requirements.txt
astroid==1.6.3
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
dj-database-url==0.5.0
django==2.0.4
djangorestframework==3.8.2
Flask==0.12.2
google-api-python-client==1.6.6
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.7.1
httplib2==0.11.3
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nodeenv==1.3.0
oauth2client==4.1.2
oauthlib==2.0.7
pipenv==11.10.0
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydrive==1.3.1
pylint==1.8.4
pytz==2018.4
pyyaml==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
whitenoise==3.3.1
wrapt==1.10.11
